fix!: Move `Name` and `NameProperty` from `UIElement` to `FrameworkElement`
